We've added an event that will reset to default zoom level on click of shelter marker. This is to ensure that the flyto offset (on mobile) will always move the same amount in the viewport of the browser. We've also increased the longitude offset to account for smaller phone screens.
In the future we may want to instead change the size of the map canvas
to some fraction of the viewport instead of using this hardcoded flyover
offset. The popup would be rendered in the remaineder of the viewport,
and the marker would just get centered in the map view.