Registration of js components in stimulus is fixed.
The regex used to match 'real' module names for registration of
ts/js components with stimulus has been extended slightly. Registration
has been wrapped in a try/except so that it doesn't break if one of the
matches is null for some reason. it just logs it to the console instead.

There was code in the shelter locations component file that removed the
search bar from the shelter map. There seemed to be some syntax that was
depricated in the statement. Something was up with it. The code line
that added captured the search bar element in a variable created a
variable that was undefined. When I commented the code out, the search
bar was still hidden. I assume it was modified somewhere else in the
codebase, so this code in redundant.